add login and registration (fix bug in authOptions)
